+++ R/distinct.R
@@ -3,6 +3,14 @@
 #' Retain only unique/distinct rows from an input tbl. This is similar
 #' to [unique.data.frame()], but considerably faster.
 #'
+#' Comparing list columns is not fully supported.
+#' Elements in list columns are compared by reference.
+#' A warning will be given when trying to include list columns in the
+#' computation.
+#' This behavior is kept for compatibility reasons and may change in a future
+#' version.
+#' See examples.
+#'
 #' @param .data a tbl
 #' @param ... Optional variables to use when determining uniqueness. If there
 #'   are multiple rows for a given combination of inputs, only the first
@@ -39,6 +47,11 @@
 #' ) %>% group_by(g)
 #' df %>% distinct()
 #' df %>% distinct(x)
+#'
+#' # Values in list columns are compared by reference, this can lead to
+#' # surprising results
+#' tibble(a = as.list(c(1, 1, 2))) %>% glimpse() %>% distinct()
+#' tibble(a = as.list(1:2)[c(1, 1, 2)]) %>% glimpse() %>% distinct()
 distinct <- function(.data, ..., .keep_all = FALSE) {
   UseMethod("distinct")
 }
@@ -61,11 +74,11 @@ distinct_vars <- function(.data, vars, group_vars = character(), .keep_all = FAL
 
   # If no input, keep all variables
   if (length(vars) == 0) {
-    list_cols_error(.data, names(.data))
+    vars <- list_cols_warning(.data, seq_along(.data))
     return(list(
       data = .data,
-      vars = names(.data),
-      keep = names(.data)
+      vars = vars,
+      keep = seq_along(.data)
     ))
   }
 
@@ -77,22 +90,31 @@ distinct_vars <- function(.data, vars, group_vars = character(), .keep_all = FAL
   out_vars <- intersect(names(.data), c(names(vars), group_vars))
 
   if (.keep_all) {
-    keep <- names(.data)
+    keep <- seq_along(.data)
   } else {
     keep <- unique(out_vars)
   }
 
-  list_cols_error(.data, keep)
+  out_vars <- list_cols_warning(.data, out_vars)
   list(data = .data, vars = out_vars, keep = keep)
 }
 
 #' Throw an error if there are tbl columns of type list
 #'
 #' @noRd
-list_cols_error <- function(df, keep_cols) {
-  if (any(map_lgl(df[keep_cols], is.list))) {
-    abort("distinct() does not support columns of type `list`")
+list_cols_warning <- function(df, keep_cols) {
+  df_keep <- df[keep_cols]
+  lists <- map_lgl(df_keep, is.list)
+  if (any(lists)) {
+    items <- fmt_items(fmt_obj(names(df_keep)[lists]))
+    warn(
+      glue("distinct() does not fully support columns of type `list`.
+            List elements are compared by reference, see ?distinct for details.
+            This affects the following columns:
+            {items}")
+    )
   }
+  keep_cols
 }
 
 #' Efficiently count the number of unique values in a set of vector