+++ tests/testthat/test-distinct.R
@@ -89,10 +89,18 @@ test_that("distinct on a new, copied variable is equivalent to mutate followed b
 test_that("distinct on a dataframe or tibble with columns of type list throws an error", {
   df <- tibble(
     a = c("1", "1", "2", "2", "3", "3"),
-    b = c(list("A"), list("A"), list("B"), list("B"), list("C"), list("C"))
+    b = list("A")
   )
-  df2 <- data.frame(x = 1:5, y = I(list(1:3, 1:3, 1:3, 1:3, 1:3)))
+  df2 <- data.frame(x = 1:5, y = I(list(1:3, 2:4, 3:5, 4:6, 5:7)))
 
-  expect_error(df %>% distinct())
-  expect_error(df2 %>% distinct())
+  expect_warning(
+    expect_identical(df %>% distinct(), df %>% slice(c(1, 3, 5))),
+    "distinct() does not fully support columns of type `list`.\nList elements are compared by reference, see ?distinct for details.\nThis affects the following columns:\n- `b`",
+    fixed = TRUE
+  )
+  expect_warning(
+    expect_identical(df2 %>% distinct(), df2),
+    "distinct() does not fully support columns of type `list`.\nList elements are compared by reference, see ?distinct for details.\nThis affects the following columns:\n- `y`",
+    fixed = TRUE
+  )
 })