+++ NEWS.md
@@ -52,7 +52,7 @@
 
 * Fix `summarise()` for empty data frames with zero columns (#3071).
 
-* Add error for `distinct()` if any of the selected columns are of type `list` (#3088, @foo-bar-baz-qux).
+* Add warning with explanation to `distinct()` if any of the selected columns are of type `list` (#3088, @foo-bar-baz-qux).
 
 * `sample_n()` and `sample_frac()` on grouped data frame are now faster especially for those with large number of groups (#3193, @saurfang).